backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
qdrant-client>=1.7.3
FlagEmbedding>=1.2.5
torch>=2.1.0
transformers>=4.36.0
scikit-learn>=1.4.0
numpy>=1.26.0
rouge-score>=0.1.2
hdbscan>=0.8.33
umap-learn>=0.5.5
pandas>=2.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
tqdm>=4.66.0
scipy>=1.11.0
